+++ src/index.ts
@@ -256,6 +256,7 @@ export type UnloadableChildNode = {
 */
 export type RenderResult = {
   nodes: ChildNode[];
+  pendingMounts: (() => void)[];
 };
 
 export type DeletedNode = {
@@ -354,10 +355,10 @@ export function createForgoInstance(customEnv: any) {
     pendingAttachStates: NodeAttachedComponentState<any>[],
     mountOnPreExistingDOM: boolean
   ): RenderResult {
-    // Array of Nodes
-    if (Array.isArray(forgoNode)) {
+    // Array of Nodes, or Fragment
+    if (Array.isArray(forgoNode) || isForgoFragment(forgoNode)) {
       return renderArray(
-        forgoNode,
+        flatten(forgoNode),
         insertionOptions,
         pendingAttachStates,
         mountOnPreExistingDOM
@@ -387,22 +388,27 @@ export function createForgoInstance(customEnv: any) {
         pendingAttachStates,
         mountOnPreExistingDOM
       );
-    } else if (isForgoFragment(forgoNode)) {
-      return renderFragment(
-        forgoNode,
-        insertionOptions,
-        pendingAttachStates,
-        mountOnPreExistingDOM
-      );
     }
     // Component.
     else {
-      return renderComponent(
+      const result = renderComponent(
         forgoNode,
         insertionOptions,
         pendingAttachStates,
         mountOnPreExistingDOM
       );
+      // In order to prevent issue #50 (Fragments having mount() called before
+      // *all* child elements have finished rendering), we delay calling mount
+      // until a subtree's render has completed
+      //
+      // Ideally this would encompass both mounts and unmounts, but an unmounted
+      // component doesn't get `renderComponent()` called on it, so we need to
+      // continue unmounting inside each of the type-specific render functions.
+      // That's fine since the problem is elements not existing at mount time,
+      // whereas unmount timing isn't sensitive to that.
+      result.pendingMounts.forEach((fn) => fn());
+      result.pendingMounts.length = 0;
+      return result;
     }
   }
 
@@ -415,7 +421,7 @@ export function createForgoInstance(customEnv: any) {
     _pendingAttachStates: NodeAttachedComponentState<any>[],
     _mountOnPreExistingDOM: boolean
   ): RenderResult {
-    return { nodes: [] };
+    return { nodes: [], pendingMounts: [] };
   }
 
   /*
@@ -472,8 +478,13 @@ export function createForgoInstance(customEnv: any) {
     }
 
     syncAttrsAndState(forgoNode, textNode, true, pendingAttachStates);
-    remountComponents(pendingAttachStates, oldComponentState);
-    return { nodes: [textNode] };
+    unmountComponents(pendingAttachStates, oldComponentState);
+    return {
+      nodes: [textNode],
+      pendingMounts: [
+        () => mountComponents(pendingAttachStates, oldComponentState),
+      ],
+    };
   }
 
   /*
@@ -639,10 +650,14 @@ export function createForgoInstance(customEnv: any) {
 
       renderChildNodes(targetElement);
       unloadMarkedNodes(targetElement, pendingAttachStates);
+      unmountComponents(pendingAttachStates, oldComponentState);
 
-      remountComponents(pendingAttachStates, oldComponentState);
-
-      return { nodes: [targetElement] };
+      return {
+        nodes: [targetElement],
+        pendingMounts: [
+          () => mountComponents(pendingAttachStates, oldComponentState),
+        ],
+      };
     }
 
     function addElement(
@@ -662,10 +677,12 @@ export function createForgoInstance(customEnv: any) {
       syncAttrsAndState(forgoElement, newElement, true, pendingAttachStates);
 
       renderChildNodes(newElement);
+      unmountComponents(pendingAttachStates, undefined);
 
-      remountComponents(pendingAttachStates, undefined);
-
-      return { nodes: [newElement] };
+      return {
+        nodes: [newElement],
+        pendingMounts: [() => mountComponents(pendingAttachStates, undefined)],
+      };
     }
   }
 
@@ -802,6 +819,7 @@ export function createForgoInstance(customEnv: any) {
             indexOfNode,
             indexOfNode + componentState.nodes.length
           ),
+          pendingMounts: [],
         };
       }
     }
@@ -989,7 +1007,7 @@ export function createForgoInstance(customEnv: any) {
         "Arrays and fragments cannot be rendered at the top level."
       );
     } else {
-      let allNodes: ChildNode[] = [];
+      const renderResults: RenderResult = { nodes: [], pendingMounts: [] };
 
       let currentNodeIndex = insertionOptions.currentNodeIndex;
       let numNodes = insertionOptions.length;
@@ -1004,63 +1022,29 @@ export function createForgoInstance(customEnv: any) {
           length: numNodes,
         };
 
-        const { nodes } = internalRender(
+        const renderResult = internalRender(
           forgoNode,
           newInsertionOptions,
           pendingAttachStates,
           mountOnPreExistingDOM
         );
 
-        allNodes = allNodes.concat(nodes);
+        renderResults.nodes.push(...renderResult.nodes);
+        renderResults.pendingMounts.push(...renderResult.pendingMounts);
 
         const totalNodesAfterRender =
           insertionOptions.parentElement.childNodes.length;
 
         const numNodesRemoved =
-          totalNodesBeforeRender + nodes.length - totalNodesAfterRender;
+          totalNodesBeforeRender +
+          renderResult.nodes.length -
+          totalNodesAfterRender;
 
-        currentNodeIndex += nodes.length;
+        currentNodeIndex += renderResult.nodes.length;
         numNodes -= numNodesRemoved;
       }
 
-      return { nodes: allNodes };
-    }
-  }
-
-  /*
-    Render a Fragment
-  */
-  function renderFragment(
-    fragment: ForgoFragment,
-    insertionOptions: NodeInsertionOptions,
-    pendingAttachStates: NodeAttachedComponentState<any>[],
-    mountOnPreExistingDOM: boolean
-  ): RenderResult {
-    return renderArray(
-      flatten(fragment),
-      insertionOptions,
-      pendingAttachStates,
-      mountOnPreExistingDOM
-    );
-  }
-
-  /**
-   * Attach component states to DOM nodes and call unmount/mount lifecycle
-   * methods
-   */
-  function remountComponents(
-    pendingAttachStates: NodeAttachedComponentState<any>[],
-    oldComponentStates: NodeAttachedComponentState<any>[] | undefined
-  ) {
-    if (oldComponentStates) {
-      const indexOfFirstIncompatibleState = findIndexOfFirstIncompatibleState(
-        pendingAttachStates,
-        oldComponentStates
-      );
-      unmountComponents(oldComponentStates, indexOfFirstIncompatibleState);
-      mountComponents(pendingAttachStates, indexOfFirstIncompatibleState);
-    } else {
-      mountComponents(pendingAttachStates, 0);
+      return renderResults;
     }
   }
 
@@ -1121,11 +1105,7 @@ export function createForgoInstance(customEnv: any) {
       if (state) {
         state.deleted = true;
         const oldComponentStates = state.components;
-        const indexOfFirstIncompatibleState = findIndexOfFirstIncompatibleState(
-          pendingAttachStates,
-          oldComponentStates
-        );
-        unmountComponents(state.components, indexOfFirstIncompatibleState);
+        unmountComponents(pendingAttachStates, oldComponentStates);
       }
     }
     clearDeletedNodes(parentElement);
@@ -1161,19 +1141,32 @@ export function createForgoInstance(customEnv: any) {
     return i;
   }
 
-  /*
-    Unmount components above an index. This is going to be passed a stale state[].
-    The from param is the index at which stale state[] differs from new state[]
-  */
+  /**
+   * Unmount components above an index. This is going to be passed a stale
+   * state[].
+   *
+   * The `unmount` lifecycle event will be called.
+   */
   function unmountComponents(
-    states: NodeAttachedComponentState<any>[],
-    from: number
+    pendingAttachStates: NodeAttachedComponentState<any>[],
+    oldComponentStates: NodeAttachedComponentState<any>[] | undefined
   ) {
+    if (!oldComponentStates) return;
+
     // If the parent has already unmounted, we can skip checks on children.
     let parentHasUnmounted = false;
 
-    for (let i = from; i < states.length; i++) {
-      const state = states[i];
+    const indexOfFirstIncompatibleState = findIndexOfFirstIncompatibleState(
+      pendingAttachStates,
+      oldComponentStates
+    );
+
+    for (
+      let i = indexOfFirstIncompatibleState;
+      i < oldComponentStates.length;
+      i++
+    ) {
+      const state = oldComponentStates[i];
       const component = state.component;
       if (component.unmount) {
         // Render if:
@@ -1208,16 +1201,27 @@ export function createForgoInstance(customEnv: any) {
     }
   }
 
-  /*
-    Mount components above an index. This is going to be passed the new state[].
-    The from param is the index at which stale state[] differs from new state[]
-  */
+  /**
+   * Mount components above an index. This is going to be passed the new
+   * state[].
+   */
   function mountComponents(
-    states: NodeAttachedComponentState<any>[],
-    from: number
+    pendingAttachStates: NodeAttachedComponentState<any>[],
+    oldComponentStates: NodeAttachedComponentState<any>[] | undefined
   ) {
-    for (let i = from; i < states.length; i++) {
-      const state = states[i];
+    const indexOfFirstIncompatibleState = oldComponentStates
+      ? findIndexOfFirstIncompatibleState(
+          pendingAttachStates,
+          oldComponentStates
+        )
+      : 0;
+
+    for (
+      let i = indexOfFirstIncompatibleState;
+      i < pendingAttachStates.length;
+      i++
+    ) {
+      const state = pendingAttachStates[i];
       if (state.component.mount && !state.isMounted) {
         state.isMounted = true;
         state.component.mount(state.props, state.args);
@@ -1723,10 +1727,11 @@ export function createForgoInstance(customEnv: any) {
               indexOfNode,
               indexOfNode + originalComponentState.nodes.length
             ),
+            pendingMounts: [],
           };
         }
       } else {
-        return { nodes: [] };
+        return { nodes: [], pendingMounts: [] };
       }
     } else {
       throw new Error(`Missing node information in rerender() argument.`);