+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
 
-- Fix #62: ensure that a child component's `mount()` lifecycle method is only called after its parent has completely finished rendering
+- Fix #62: ensure that a child component's `mount` lifecycle method is only
+  called after its parent has completely finished rendering.
+- Fix #50: Components that returned a fragment saw their `mount` lifecycle
+  method called after the first child element had been created instead of after
+  the render had completed. 
 
 # 3.1.1
 - Fix: components that changed their root HTML tag were erroneously unmounted